+++ packages/flutter_gemma_speech/test/tts_fail_loud_test.dart
@@ -1,10 +1,10 @@
 import 'dart:typed_data';
 
-import 'package:flutter_gemma_speech/src/tts/tts_text_frontend.dart';
+import 'package:flutter_gemma_speech/src/tts/matcha_text_frontend.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
-  final fe = TtsTextFrontend(
+  final fe = MatchaTextFrontend(
     symbolToId: {'_': 0, 'a': 1},
     dictionary: {'x': 'aq'}, // 'q' not in table
     embeddingTable: Float32List.fromList([0, 0, 10, 11]),
