+++ packages/flutter_gemma_speech/test/tts_text_frontend_test.dart
@@ -1,12 +1,12 @@
 import 'dart:typed_data';
 
+import 'package:flutter_gemma_speech/src/tts/matcha_text_frontend.dart';
 import 'package:flutter_gemma_speech/src/tts/tts_frontend_input.dart';
-import 'package:flutter_gemma_speech/src/tts/tts_text_frontend.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
   // Tiny 4-symbol table: '_'(0 blank/pad), 'a'(1), 'b'(2), '.'(3); 2 channels.
-  final frontend = TtsTextFrontend(
+  final frontend = MatchaTextFrontend(
     symbolToId: {'_': 0, 'a': 1, 'b': 2, '.': 3},
     dictionary: {'ab': 'ab'}, // word "ab" -> IPA "ab"
     embeddingTable: Float32List.fromList([
@@ -32,6 +32,10 @@ void main() {
     expect(out.symbolEmbeddings.sublist(2, 4), [10, 11]);
     expect(out.symbolEmbeddings.sublist(6, 8), [20, 21]);
     expect(out.symbolEmbeddings.sublist(0, 2), [0, 0]); // blank row
+  });
+
+  test('TtsTextFrontend.encode returns MatchaFrontendInput', () {
+    expect(frontend.encode('ab.'), isA<MatchaFrontendInput>());
   });
 
   test('OOV word throws (dictionary-only)', () {
